% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverscarial.r
\name{advGridMinChange}
\alias{advGridMinChange}
\title{Grid search of min change adversarial attack. Tries each
combination on a cluster, given a list of genes and a list of modifications.}
\usage{
advGridMinChange(
  exprs,
  clusters,
  target,
  classifier,
  genes,
  modifications = list(list("perc1"), list("perc99")),
  return_first_found = FALSE,
  verbose = FALSE,
  iamsure = FALSE
)
}
\arguments{
\item{exprs}{a matrix or dataframe of numeric RNA expression,
cells are rows and genes are columns.}

\item{clusters}{a list of the clusters to which the cells belong}

\item{target}{the name of the cluster to modify}

\item{classifier}{a classifier in the suitable format}

\item{genes}{the list of genes to study}

\item{modifications}{the list of the modifications to study}

\item{return_first_found}{set to TRUE to return result when a
the first misclassification is found}

\item{verbose}{logical, set to TRUE to activate verbose mode}

\item{iamsure}{logical, prevents from expansive calculations
when \code{genes} list is too long, set to \code{TRUE} to run anyway.}
}
\value{
dataframe results of the classification of all the grid combinations
}
\description{
Grid search of min change adversarial attack. Tries each
combination on a cluster, given a list of genes and a list of modifications.
}
\examples{
MyClassifier <- function(expr, clusters, target) {
   c("T cell", 0.9)
}
rna_expression <- data.frame(CD4=c(0,0,0,0), CD8A=c(1,1,1,1),
     CD8B=c(2,2,3,3))
genes <- c("CD4", "CD8A")
clusters_id <- c("B cell","B cell","T cell","T cell")

advGridMinChange(rna_expression, clusters_id, "T cell",
MyClassifier, genes=genes,
modifications = list(list("perc1"), list("perc99")))
}
