% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverscarial.r
\name{advMaxChange}
\alias{advMaxChange}
\title{Find a max change adversarial attack. It finds the longer
list of genes you can modify on a cluster without changing its
classification.}
\usage{
advMaxChange(
  exprs,
  clusters,
  target,
  classifier,
  excl_genes = c(),
  genes = c(),
  adv_method = "fixed",
  adv_fixed_value = 3,
  adv_fct = NULL,
  max_split_size = 1,
  verbose = FALSE
)
}
\arguments{
\item{exprs}{a matrix or dataframe of numeric RNA expression,
cells are rows and genes are columns.}

\item{clusters}{a list of the clusters to which the cells belong}

\item{target}{the name of the cluster to modify}

\item{classifier}{a classifier in the suitable format}

\item{excl_genes}{a list of genes to exclude from the analysis}

\item{genes}{a list of genes in case you want to limit the
attack on a subset of genes}

\item{adv_method}{the name of the method to use}

\item{adv_fixed_value}{the numeric value to use in case of
adv_method=\code{fixed}}

\item{adv_fct}{the function to use in case adv_method
belongs to the following list: \code{full_row_fct}, \code{target_row_fct},
\code{target_matrix_fct}, \code{full_matrix_fct}}

\item{max_split_size}{max size of dichotomic slices.}

\item{verbose}{logical, set to TRUE to activate verbose mode}
}
\value{
a list of genes you can modify on a cluster without
modifying its classification
}
\description{
Find a max change adversarial attack. It finds the longer
list of genes you can modify on a cluster without changing its
classification.
}
\examples{
advMaxChange(rna_expression, clusters_id,
"Memory CD4+", myClassifier, adv_method="perc99")
}
