% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverscarial.r
\name{advMinChange}
\alias{advMinChange}
\title{Find a one gene min change adversarial attack list. A one gene min change
adversarial attack  refers to the modification of a single gene
within a cluster, leading to a change in its classification. The function
returns a list of genes/new classification.}
\usage{
advMinChange(
  exprs,
  clusters,
  target,
  classifier,
  excl_genes = c(),
  genes = c(),
  adv_method = "perc99",
  adv_fixed_value = 3,
  adv_fct = NULL,
  first_dichot = 100,
  max_split_size = 1,
  return_first_found = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{exprs}{a matrix or dataframe of numeric RNA expression,
cells are rows and genes are columns.}

\item{clusters}{a list of the clusters to which the cells belong}

\item{target}{the name of the cluster to modify}

\item{classifier}{a classifier in the suitable format}

\item{excl_genes}{a list of genes to exclude from the analysis}

\item{genes}{a list of genes in case you want to limit the
attack on a subset of genes}

\item{adv_method}{the name of the method to use}

\item{adv_fixed_value}{the numeric value to use in case of
adv_method=\code{fixed}}

\item{adv_fct}{the function to use in case adv_method
belongs to the following list: \code{full_row_fct}, \code{target_row_fct},
\code{target_matrix_fct}, \code{full_matrix_fct}}

\item{first_dichot}{the initial number of slices before
the dichotomic search}

\item{max_split_size}{max size of dichotomic slices}

\item{return_first_found}{set to TRUE to return result when a
the first misclassification is found}

\item{verbose}{logical, set to TRUE to activate verbose mode}
}
\value{
a list of genes/new classification tuples
}
\description{
Find a one gene min change adversarial attack list. A one gene min change
adversarial attack  refers to the modification of a single gene
within a cluster, leading to a change in its classification. The function
returns a list of genes/new classification.
}
\examples{
advMinChange(rna_expression, clusters_id,
"Memory CD4+", myClassifier, adv_method="perc99")
}
