% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverscarial.r
\name{advModifications}
\alias{advModifications}
\title{Returns a modified RNA expression matrix, for a given cluster,
for a given modification.}
\usage{
advModifications(
  exprs,
  genes,
  clusters,
  target,
  adv_method = "fixed",
  adv_fixed_value = 3,
  adv_fct = NULL,
  verbose = FALSE
)
}
\arguments{
\item{exprs}{a matrix or dataframe of numeric RNA expression,
cells are rows and genes are columns.}

\item{genes}{the list of genes to modify}

\item{clusters}{a list of the clusters to which the cells belong}

\item{target}{the name of the cluster to modify}

\item{adv_method}{the name of the method to use}

\item{adv_fixed_value}{the numeric value to use in case of
adv_method=\code{fixed}}

\item{adv_fct}{the function to use in case adv_method
belongs to the following list: \code{full_row_fct}, \code{target_row_fct},
\code{target_matrix_fct}, \code{full_matrix_fct}}

\item{verbose}{logical, set to TRUE to activate verbose mode}
}
\value{
the matrix or a dataframe exprs modified on asked genes
with the specified modification
}
\description{
Returns a modified RNA expression matrix, for a given cluster,
for a given modification.
}
\examples{
advModifications(rna_expression, c("CD4", "CD8"), clusters_id,
"Memory CD4+", adv_method="perc99")
}
