% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverscarial.r
\name{advRandWalkMinChange}
\alias{advRandWalkMinChange}
\title{Random walk search of min change adversarial attack.
Step 1 is to find a seed by trying random combinations of
genes and modifications on a cluster until the classification is altered.
Step 2 is to perform a random walk search to reduce the number of genes
needed to change the classification."}
\usage{
advRandWalkMinChange(
  exprs,
  clusters,
  target,
  classifier,
  genes,
  modifications = list(list("perc1"), list("perc99")),
  first_batch = 100,
  walk_length = 100,
  step_change_ratio = 0.2,
  while_max_count = 10000,
  verbose = FALSE
)
}
\arguments{
\item{exprs}{a matrix or dataframe of numeric RNA expression,
cells are rows and genes are columns.}

\item{clusters}{a list of the clusters to which the cells belong}

\item{target}{the name of the cluster to modify}

\item{classifier}{a classifier in the suitable format}

\item{genes}{the list of genes to study}

\item{modifications}{the list of the modifications to study}

\item{first_batch}{the maximum number of try in step 1}

\item{walk_length}{the maximum number of try in step 2}

\item{step_change_ratio}{ratio of parameters change in new walk step}

\item{while_max_count}{the maximum number of try when looking
for new combination of parameters}

\item{verbose}{logical, set to TRUE to activate verbose mode}
}
\value{
dataframe results of the classification of all the grid combinations
}
\description{
Random walk search of min change adversarial attack.
Step 1 is to find a seed by trying random combinations of
genes and modifications on a cluster until the classification is altered.
Step 2 is to perform a random walk search to reduce the number of genes
needed to change the classification."
}
\examples{
advRandWalkMinChange(rna_expression, clusters_id, "Memory CD4+",
myClassifier, genes=c("CD4","CD8","IL2A","IL2B"),
modifications = list(list("perc1"), list("perc99")))

# Stop at first attack discovery, whitout going into the walk
# parameter search.
advRandWalkMinChange(rna_expression, clusters_id, "Memory CD4+",
myClassifier, genes=c("CD4","CD8","IL2A","IL2B"),
modifications = list(list("perc1"), list("perc99")), walk_length=0)
}
