% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{matrixFromSCE}
\alias{matrixFromSCE}
\title{Returns the RNA expression matrix from a SingleCellExperiment
with unique hgnc gene names in columns}
\usage{
matrixFromSCE(sce)
}
\arguments{
\item{sce}{SingleCellExperiment object to convert}
}
\value{
the RNA expression matrix from a SingleCellExperiment
with unique hgnc gene names in columns
}
\description{
Returns the RNA expression matrix from a SingleCellExperiment
with unique hgnc gene names in columns
}
\examples{
library(TENxPBMCData)

pbmc <- TENxPBMCData(dataset = "pbmc3k")
mat_rna <- matrixFromSCE(pbmc)

}
