% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.r
\name{MClassifier}
\alias{MClassifier}
\title{Example cell type classifier for the pbmc3k dataset}
\usage{
MClassifier(expr, clusters, target)
}
\arguments{
\item{expr}{RNA expression matrix}

\item{clusters}{list of clusters to which each cell belongs}

\item{target}{name of the cell cluster to classify}
}
\value{
a vector with the classification, and the odd
}
\description{
Example cell type classifier for the pbmc3k dataset
}
\examples{
library(TENxPBMCData)

pbmc <- TENxPBMCData(dataset = "pbmc3k")
mat_rna <- matrixFromSCE(pbmc)
cell_types <- system.file("extdata",
    "pbmc3k_cell_types.tsv",
    package = "adverSCarial"
)
cell_types <- read.table(cell_types, sep = "\t")$cell_type

MClassifier(mat_rna, cell_types, "DC")

}
