% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverscarial.r
\name{minChangeOverview}
\alias{minChangeOverview}
\title{Run an approximation of advMinChange on all clusters
for a given modification. The precision of the
approximation can be controlled using the \code{max_split_size}
parameter, with lower values resulting in greater precision
but longer processing time. The purpose of this
approximation is to determine which clusters are most
susceptible to min change adversarial attacks.}
\usage{
minChangeOverview(
  exprs,
  clusters,
  classifier,
  excl_genes = c(),
  genes = c(),
  modifications = list(list("perc1"), list("perc99")),
  adv_method = "fixed",
  adv_fixed_value = 3,
  adv_fct = NULL,
  first_dichot = 100,
  max_split_size = 100,
  verbose = FALSE
)
}
\arguments{
\item{exprs}{a matrix or dataframe of numeric RNA expression,
cells are rows and genes are columns.}

\item{clusters}{a list of the clusters to which the cells belong}

\item{classifier}{a classifier in the suitable format}

\item{excl_genes}{a list of genes to exclude from the analysis}

\item{genes}{a list of genes in case you want to limit the
analysis on a subset of genes}

\item{modifications}{the list of the modifications to study}

\item{adv_method}{the name of the method to use}

\item{adv_fixed_value}{the numeric value to use in case of
adv_method=\code{fixed}}

\item{adv_fct}{the function to use in case adv_method
belongs to the following list: \code{full_row_fct}, \code{target_row_fct},
\code{target_matrix_fct}, \code{full_matrix_fct}}

\item{first_dichot}{the initial number of slices before
the dichotomic search}

\item{max_split_size}{max size of dichotomic slices.}

\item{verbose}{logical, set to TRUE to activate verbose mode}
}
\value{
a list of genes/new classification tuples
}
\description{
Run an approximation of advMinChange on all clusters
for a given modification. The precision of the
approximation can be controlled using the \code{max_split_size}
parameter, with lower values resulting in greater precision
but longer processing time. The purpose of this
approximation is to determine which clusters are most
susceptible to min change adversarial attacks.
}
\examples{
minChangeOverview(rna_expression, clusters_id,
myClassifier, modifications = list(list("perc1"), list("perc99")))
}
