% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverscarial.r
\name{predictWithNewValue}
\alias{predictWithNewValue}
\title{Returns a classification and an odd value from a
RNA expression matrix, for given genes, for a given cluster,
for a given modification.}
\usage{
predictWithNewValue(
  exprs,
  genes,
  clusters,
  target,
  classifier,
  adv_method = "perc99",
  adv_fixed_value = 3,
  adv_fct = NULL,
  verbose = FALSE
)
}
\arguments{
\item{exprs}{a matrix or dataframe of numeric RNA expression,
cells are rows and genes are columns.}

\item{genes}{the list of genes to modify}

\item{clusters}{a list of the clusters to which the cells belong}

\item{target}{the name of the cluster to modify}

\item{classifier}{a classifier in the suitable format}

\item{adv_method}{the name of the method to use}

\item{adv_fixed_value}{the numeric value to use in case of
adv_method=\code{fixed}}

\item{adv_fct}{the function to use in case adv_method
belongs to the following list: \code{full_row_fct}, \code{target_row_fct},
\code{target_matrix_fct}, \code{full_matrix_fct}}

\item{verbose}{logical, set to TRUE to activate verbose mode}
}
\value{
a vector of the classification, and the associated odd
}
\description{
Returns a classification and an odd value from a
RNA expression matrix, for given genes, for a given cluster,
for a given modification.
}
\examples{
MyClassifier <- function(expr, clusters, target) {
   c("T cell", 0.9)
}
rna_expression <- data.frame(CD4=c(0,0,0,0), CD8A=c(1,1,1,1),
     CD8B=c(2,2,3,3))
genes <- c("CD4", "CD8A")
clusters_id <- c("B cell","B cell","T cell","T cell")

predictWithNewValue(rna_expression, genes, clusters_id,
"T cell", MyClassifier, adv_method="perc99")
}
